refactor new tests as per agreed file structure
