+++ arc_application/tests/test_upload_capita_dbs/test_routing.py
@@ -1,17 +1,15 @@
 import json
 from unittest import mock
 
-from ..models import CapitaDBSFile
-from ..services import dbs_api
-from ..views.base import custom_login
-from ..views import __handle_file_upload
+from ...models import CapitaDBSFile
+from ...services import dbs_api
+from ...views.base import custom_login
+from ...views import __handle_file_upload
 
 from django.conf import settings
 from django.contrib.auth.models import Group, User
-from django.db import InternalError
-from django.forms import ValidationError
 from django.http import HttpResponse
-from django.test import RequestFactory, SimpleTestCase, TestCase
+from django.test import TestCase
 from django.urls import resolve, reverse
 
 
@@ -86,7 +84,9 @@ class UploadCapitaDBSRoutingTests(TestCase):
         csv_file.close()
 
         self.assertEqual(response.status_code, 200)
-        self.assertFormError(response, 'form', field='capita_list_file', errors='There was an error with the file you tried to upload. Check the file and try again')
+        self.assertFormError(response, 'form', field='capita_list_file',
+                             errors='There was an error with the file you tried to upload. '
+                             'Check the file and try again')
 
     def test_error_added_to_form_if_not_400_or_201_status_code_from_dbs_api(self, dbs_api_mock):
         dbs_api_mock.return_value.status_code = 500
@@ -98,69 +98,7 @@ class UploadCapitaDBSRoutingTests(TestCase):
         csv_file.close()
 
         self.assertEqual(response.status_code, 200)
-        self.assertFormError(response, 'form', field='capita_list_file', errors='We couldn\'t upload your new list. Try again')
+        self.assertFormError(response, 'form', field='capita_list_file',
+                             errors='We couldn\'t upload your new list. Try again')
 
 
-@mock.patch.object(dbs_api, 'batch_overwrite', return_value=HttpResponse(status=201))
-class UploadCapitaDBSHelperFunctionTests(SimpleTestCase):
-    def setUp(self):
-        self.factory = RequestFactory()
-
-    def test_formatting_of_previous_upload_information(self, dbs_api_mock):
-        self.skipTest('testNotImplemented')
-
-    def test_validation_error_raised_if_400_status_code_from_dbs_api(self, dbs_api_mock):
-        dbs_api_mock.return_value.status_code = 400
-        dbs_api_mock.return_value.text = json.dumps('Test error message')
-
-        with open('arc_application/tests/resources/test_csv.csv') as csv_file:
-            request = self.factory.post(reverse('Upload-Capita-DBS'), {'capita_list_file': csv_file})
-            request_files = request.FILES
-
-        csv_file.close()
-
-        with self.assertRaisesMessage(ValidationError, 'Test error message'):
-            handle_file_upload(request_files)
-
-    def test_internal_error_raised_if_not_201_or_400_status_code_from_dbs_api(self, dbs_api_mock):
-        dbs_api_mock.return_value.status_code = 500
-        dbs_api_mock.return_value.text = json.dumps('Test error message')
-
-        with open('arc_application/tests/resources/test_csv.csv') as csv_file:
-            request = self.factory.post(reverse('Upload-Capita-DBS'), {'capita_list_file': csv_file})
-            request_files = request.FILES
-
-        csv_file.close()
-
-        with self.assertRaisesMessage(InternalError, 'The DBS API returned a 500 status code. Response text: Test error message'):
-            handle_file_upload(request_files)
-
-    def test_no_error_raised_if_201_status_code_from_dbs_api(self, dbs_api_mock):
-        dbs_api_mock.return_value.status_code = 201
-
-        with open('arc_application/tests/resources/test_csv.csv') as csv_file:
-            request = self.factory.post(reverse('Upload-Capita-DBS'), {'capita_list_file': csv_file})
-            request_files = request.FILES
-
-        csv_file.close()
-
-        x = handle_file_upload(request_files)
-
-        self.assertEqual(x, None)
-
-
-class UploadCapitaDBSFormTests(SimpleTestCase):
-    def setUp(self):
-        self.factory = RequestFactory()
-
-    def test_no_file_selected_raises_error(self):
-        self.skipTest('testNotImplemented')
-
-    def test_invalid_file_extension_raises_error(self):
-        self.skipTest('testNotImplemented')
-
-    def test_csvx_file_passes_validation(self):
-        self.skipTest('testNotImplemented')
-
-    def test_csv_file_passes_validation(self):
-        self.skipTest('testNotImplemented')
